Rename crypto_parameters -> crypto_constants
